fixed dashboard, added logo, addded title, made alumni page server side, improved alumni cards
